Tests for running Python on Flink.

+++ sdks/python/apache_beam/runners/portability/portable_runner_test.py
@@ -82,19 +82,12 @@ class PortableRunnerTest(fn_api_runner_test.FnApiRunnerTest):
 
   @classmethod
   def _start_local_runner_subprocess_job_service(cls):
-    if cls._subprocess:
-      # Kill the old one if it exists.
-      cls._subprocess.kill()
+    cls._maybe_kill_subprocess()
     # TODO(robertwb): Consider letting the subprocess pick one and
     # communicate it back...
     port = cls._pick_unused_port()
     logging.info('Starting server on port %d.', port)
-    cls._subprocess = subprocess.Popen([
-        sys.executable, '-m',
-        'apache_beam.runners.portability.local_job_service_main', '-p',
-        str(port), '--worker_command_line',
-        '%s -m apache_beam.runners.worker.sdk_worker_main' % sys.executable
-    ])
+    cls._subprocess = subprocess.Popen(cls._subprocess_command(port))
     address = 'localhost:%d' % port
     job_service = beam_job_api_pb2_grpc.JobServiceStub(
         grpc.insecure_channel(address))
@@ -116,7 +109,7 @@ class PortableRunnerTest(fn_api_runner_test.FnApiRunnerTest):
               beam_job_api_pb2.GetJobStateRequest(job_id='[fake]'))
           break
         except grpc.RpcError as exn:
-          if exn.code != grpc.StatusCode.UNAVAILABLE:
+          if exn.code() != grpc.StatusCode.UNAVAILABLE:
             # We were able to contact the service for our fake state request.
             break
     logging.info('Server ready.')
@@ -147,14 +140,21 @@ class PortableRunnerTest(fn_api_runner_test.FnApiRunnerTest):
 
   @classmethod
   def tearDownClass(cls):
-    if hasattr(cls, '_subprocess'):
+    cls._maybe_kill_subprocess()
+
+  @classmethod
+  def _maybe_kill_subprocess(cls):
+    if hasattr(cls, '_subprocess') and cls._subprocess.poll() is None:
       cls._subprocess.kill()
       time.sleep(0.1)
 
-  def create_pipeline(self):
+  def create_options(self):
     options = PipelineOptions()
     options.view_as(PortableOptions).job_endpoint = self._get_job_endpoint()
-    return beam.Pipeline(self.get_runner(), options)
+    return options
+
+  def create_pipeline(self):
+    return beam.Pipeline(self.get_runner(), self.create_options())
 
   def test_assert_that(self):
     # TODO: figure out a way for runner to parse and raise the
@@ -194,6 +194,16 @@ class PortableRunnerTestWithGrpc(PortableRunnerTest):
 class PortableRunnerTestWithSubprocesses(PortableRunnerTest):
   _use_grpc = True
   _use_subprocesses = True
+
+  @classmethod
+  def _subprocess_command(cls, port):
+    return [
+        sys.executable,
+        '-m', 'apache_beam.runners.portability.local_job_service_main',
+        '-p', str(port),
+        '--worker_command_line',
+        '%s -m apache_beam.runners.worker.sdk_worker_main' % sys.executable,
+    ]
 
 
 if __name__ == '__main__':
